+++ TeamProject_Malang/src/main/java/com/malcom/malang/controller/HomeController.java
@@ -1,0 +1,24 @@
+package com.malcom.malang.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+
+import lombok.extern.slf4j.Slf4j;
+
+@Slf4j
+@Controller
+public class HomeController {
+
+	@RequestMapping(value = "/", method = RequestMethod.GET)
+	public String home() {
+		return "home";
+	}		
+	
+	@RequestMapping(value="write", method=RequestMethod.GET)
+	public String write() {
+		
+		return "write";
+	}
+	
+}
