+++ TeamProject_Malang/src/main/webapp/WEB-INF/spring/appServlet/servlet-context.xml
@@ -21,7 +21,8 @@
 		<beans:property name="suffix" value=".jsp" />
 	</beans:bean>
 	
-	<context:component-scan base-package="com.malcom.malang" />
+	<context:component-scan base-package="com.malcom.malang.controller" />
+	<context:component-scan base-package="com.malcom.malang.service.impl" />
